Changed ComputeNextT so it is only called for external steps when sing EQUIPMENT_FIRST_STEP.
